Fix wrong path-to-stl in xml files & Use relative path
Since the xml files always go with the cassie-stl-meshes folder in the same directory, I want to use relative path, i.e. just 'cassie-stl-meshes' in the  meshdir  property of the  <compiler>  tag in the xml files. This helps reduce the path name complications, but won't work if the xml files are not in the same directory as the cassie-stl-meshes folder.